Updated README
I've updated the code example as a [few users](https://github.com/negomi/react-burger-menu/issues/183), myself included, blindly copied and pasted the example to test out the library and were faced with an error. It wasn't clear to me at first glance that 'nameOfAnimation' needed to be changed. Hopefully my modification will make it clear for any other user who stumbles on this.